attendance_system/requirements.txt
tensorflow==2.6.0
numpy==1.19.5
scipy==1.7.1
opencv-python==4.5.3.56
Flask==2.0.1
mysql-connector-python==8.0.26
python-dotenv==0.19.0